ast: make some simple data types `Copy`
This changes some of the simpler enum's and struct's derives in the ast
to include `Copy` when implementing it is possible. This is a
convenience when using the library, as you can just use e.g.
`BinaryOperator` like you would a plain value, instead of having to use
`binop.clone()`.